[lldb] Merge SwiftLanguageRuntimeImpl back into SwiftLanguageRuntime (NFC)
The lazy initialization of the Runtime used to be necessary before the
ReflectionContext was able to initialized lazily. Since we can just
defer initializing ReflectionContext, we can get rid of the extra
indirection.
